Separate business logic from command logic for ansible role
